Remove unused REST endpoints and tRPC procedures
Audit of the backend HTTP + tRPC surface against all in-repo clients
(web, electron, mobile, CLI) found endpoints with no caller anywhere.

REST routes removed:
- GET  /api/node/metadata (legacy alias of /api/nodes/metadata)
- GET  /api/files/download (file browser uses tRPC + /api/files/local)
- GET  /api/jobs/triggers/running, POST /api/jobs/triggers/:id/{start,stop}
       (standalone stubs; whole routes/jobs.ts plugin dropped)
- POST /api/kie/webhook/:taskId (per-task variant never constructed;
       generic /api/kie/webhook stays)

Dead handlers pruned from http-api.ts (trigger stubs) and file-api.ts
(handleDownload + resolveSandboxed).

tRPC procedures removed (no client caller): the skills router (fonts
router kept, file renamed to fonts.ts), timeline.versions.* and
timeline.clips.{delete,duplicate}, plus dead procedures across assets,
collections, costs, files, jobs, messages, models, nodes, sketch,
storage, threads, users, worker, workflows, and workspace routers.

Now-unused in-file schemas, imports, and helpers cleaned up; matching
test blocks removed. Kept externally-consumed endpoints (OAuth
callbacks, kie webhook, /v1 OpenAI-compatible API, public workflow
routes, health probes, /mcp) and public procedures that may serve
out-of-repo consumers.
