+++ packages/protocol/src/api-schemas/fonts.ts
@@ -1,8 +1,8 @@
 import { z } from "zod";
 
 // ── Fonts list response ──────────────────────────────────────────
-// Mirrors `handleFontsRequest` in skills-api.ts — returns a sorted unique
-// list of system font family names (no file extensions).
+// The fonts tRPC router returns a sorted unique list of system font family
+// names (no file extensions).
 export const listOutput = z.object({
   fonts: z.array(z.string())
 });
