+++ packages/websocket/src/trpc/router.ts
@@ -14,7 +14,7 @@ import { packsRouter } from "./routers/packs.js";
 import { sandboxesRouter } from "./routers/sandboxes.js";
 import { scriptsRouter } from "./routers/scripts.js";
 import { settingsRouter } from "./routers/settings.js";
-import { skillsRouter, fontsRouter } from "./routers/skills.js";
+import { fontsRouter } from "./routers/fonts.js";
 import { storageRouter } from "./routers/storage.js";
 import { threadsRouter } from "./routers/threads.js";
 import { sketchRouter } from "./routers/sketch.js";
@@ -46,7 +46,6 @@ export const appRouter = router({
   settings: settingsRouter,
   sketch: sketchRouter,
   storyboards: storyboardsRouter,
-  skills: skillsRouter,
   storage: storageRouter,
   threads: threadsRouter,
   timeline: timelineRouter,
