+++ packages/websocket/tests/file-api.test.ts
@@ -1,5 +1,5 @@
 /**
- * Tests for file-api.ts — binary download only.
+ * Tests for file-api.ts — local-file streaming.
  *
  * JSON ops (list, info) have been migrated to the tRPC `files` router.
  * See trpc-files.test.ts for the tRPC unit tests.
@@ -27,116 +27,6 @@ afterEach(async () => {
 function makeRequest(urlPath: string, method = "GET"): Request {
   return new Request(`http://localhost${urlPath}`, { method });
 }
-
-// ---------------------------------------------------------------------------
-// Production guard
-// ---------------------------------------------------------------------------
-
-describe("production guard", () => {
-  it("returns 403 in production", async () => {
-    process.env["NODETOOL_ENV"] = "production";
-    const res = await handleFileRequest(
-      makeRequest("/api/files/download?path=test.txt")
-    );
-    expect(res.status).toBe(403);
-    const body = await res.json();
-    expect(body.detail).toContain("production");
-  });
-});
-
-// ---------------------------------------------------------------------------
-// Method restriction
-// ---------------------------------------------------------------------------
-
-describe("method restriction", () => {
-  it("rejects non-GET methods", async () => {
-    const res = await handleFileRequest(
-      makeRequest("/api/files/download?path=test.txt", "POST"),
-      { rootDir: tmpDir }
-    );
-    expect(res.status).toBe(405);
-  });
-});
-
-// ---------------------------------------------------------------------------
-// Path traversal protection (for download)
-// ---------------------------------------------------------------------------
-
-describe("path traversal protection", () => {
-  it("blocks ../../../etc/passwd traversal", async () => {
-    const res = await handleFileRequest(
-      makeRequest("/api/files/download?path=../../../etc/passwd"),
-      { rootDir: tmpDir }
-    );
-    expect(res.status).toBe(403);
-    const body = await res.json();
-    expect(body.detail).toContain("traversal");
-  });
-
-  it("blocks absolute path outside root", async () => {
-    const res = await handleFileRequest(
-      makeRequest("/api/files/download?path=/etc/passwd"),
-      { rootDir: tmpDir }
-    );
-    // The sandbox resolution should catch this
-    expect([403, 404]).toContain(res.status);
-  });
-
-  it("blocks a symlink under the root that points outside it", async () => {
-    // A symlink living inside the sandbox but resolving to an external target
-    // passes a lexical prefix check yet must be rejected after realpath.
-    const secret = path.join(os.tmpdir(), "file-api-secret.txt");
-    await fs.writeFile(secret, "top secret");
-    const link = path.join(tmpDir, "escape.txt");
-    await fs.symlink(secret, link);
-
-    const res = await handleFileRequest(
-      makeRequest("/api/files/download?path=escape.txt"),
-      { rootDir: tmpDir }
-    );
-    await fs.rm(secret, { force: true });
-    expect(res.status).toBe(403);
-    const body = await res.json();
-    expect(body.detail).toContain("traversal");
-  });
-});
-
-// ---------------------------------------------------------------------------
-// /api/files/download (binary)
-// ---------------------------------------------------------------------------
-
-describe("/api/files/download", () => {
-  it("downloads file content", async () => {
-    await fs.writeFile(path.join(tmpDir, "hello.txt"), "hello world");
-
-    const res = await handleFileRequest(
-      makeRequest("/api/files/download?path=hello.txt"),
-      { rootDir: tmpDir }
-    );
-    expect(res.status).toBe(200);
-    expect(res.headers.get("content-type")).toBe("application/octet-stream");
-    const text = await res.text();
-    expect(text).toBe("hello world");
-  });
-
-  it("rejects directory download", async () => {
-    await fs.mkdir(path.join(tmpDir, "adir"));
-
-    const res = await handleFileRequest(
-      makeRequest("/api/files/download?path=adir"),
-      { rootDir: tmpDir }
-    );
-    expect(res.status).toBe(400);
-  });
-
-  it("returns 404 for non-existent file", async () => {
-    const res = await handleFileRequest(
-      makeRequest("/api/files/download?path=missing.txt"),
-      { rootDir: tmpDir }
-    );
-    expect(res.status).toBe(404);
-  });
-});
 
 // ---------------------------------------------------------------------------
 // /api/files/local (streaming by absolute path)
