+++ packages/websocket/tests/trpc-files.test.ts
@@ -154,68 +154,4 @@ describe("files router", () => {
       });
     });
   });
-
-  // ── info ──────────────────────────────────────────────────────────────────
-
-  describe("info", () => {
-    it("returns metadata for a single file", async () => {
-      (fsPromises.stat as ReturnType<typeof vi.fn>).mockResolvedValue({
-        size: 512,
-        isDirectory: () => false,
-        mtime: new Date("2026-04-01T10:00:00.000Z")
-      });
-      const caller = createCaller(makeCtx());
-      const result = await caller.files.info({ path: "myfile.txt" });
-      expect(result).toMatchObject({
-        name: "myfile.txt",
-        path: path.join(HOME, "myfile.txt"),
-        size: 512,
-        is_dir: false
-      });
-      expect(result.modified_at).toBe("2026-04-01T10:00:00.000Z");
-    });
-
-    it("returns metadata for a directory", async () => {
-      (fsPromises.stat as ReturnType<typeof vi.fn>).mockResolvedValue({
-        size: 4096,
-        isDirectory: () => true,
-        mtime: new Date("2026-02-01T00:00:00.000Z")
-      });
-      const caller = createCaller(makeCtx());
-      const result = await caller.files.info({ path: "mydir" });
-      expect(result.is_dir).toBe(true);
-    });
-
-    it("throws NOT_FOUND when file does not exist", async () => {
-      (fsPromises.stat as ReturnType<typeof vi.fn>).mockRejectedValue(
-        new Error("ENOENT: no such file or directory")
-      );
-      const caller = createCaller(makeCtx());
-      await expect(
-        caller.files.info({ path: "missing.txt" })
-      ).rejects.toMatchObject({ code: "NOT_FOUND" });
-    });
-
-    it("throws FORBIDDEN for path traversal", async () => {
-      const caller = createCaller(makeCtx());
-      await expect(
-        caller.files.info({ path: "../../etc/passwd" })
-      ).rejects.toMatchObject({ code: "FORBIDDEN" });
-    });
-
-    it("throws FORBIDDEN in production mode", async () => {
-      process.env["NODETOOL_ENV"] = "production";
-      const caller = createCaller(makeCtx());
-      await expect(caller.files.info({ path: "." })).rejects.toMatchObject({
-        code: "FORBIDDEN"
-      });
-    });
-
-    it("rejects unauthenticated callers", async () => {
-      const caller = createCaller(makeCtx({ userId: null }));
-      await expect(caller.files.info({ path: "." })).rejects.toMatchObject({
-        code: "UNAUTHORIZED"
-      });
-    });
-  });
 });
