+++ packages/websocket/tests/trpc-timeline-coverage.test.ts
@@ -389,44 +389,3 @@ describe("timeline.clips.create", () => {
     ).rejects.toThrow(/media output/i);
   });
 });
-
-describe("timeline auth branches", () => {
-  beforeEach(() => vi.resetAllMocks());
-  afterEach(() => vi.restoreAllMocks());
-
-  it("versions.list rejects unauthenticated callers before any DB read", async () => {
-    const caller = createCaller(makeCtx({ userId: null }));
-    await expect(
-      caller.timeline.versions.list({ id: "seq-1", clipId: "c" })
-    ).rejects.toThrow();
-    expect(TS.findById).not.toHaveBeenCalled();
-  });
-
-  it("clips.duplicate 404s on unknown clip id", async () => {
-    TS.findById.mockResolvedValue(
-      makeSeq({
-        document: JSON.stringify({ tracks: [], clips: [], markers: [] })
-      })
-    );
-    const caller = createCaller(makeCtx());
-    await expect(
-      caller.timeline.clips.duplicate({
-        id: "seq-1",
-        clipId: "missing",
-        deltaMs: 0
-      })
-    ).rejects.toThrow(/clip not found/i);
-  });
-
-  it("clips.delete 404s on unknown clip id", async () => {
-    TS.findById.mockResolvedValue(
-      makeSeq({
-        document: JSON.stringify({ tracks: [], clips: [], markers: [] })
-      })
-    );
-    const caller = createCaller(makeCtx());
-    await expect(
-      caller.timeline.clips.delete({ id: "seq-1", clipId: "missing" })
-    ).rejects.toThrow(/clip not found/i);
-  });
-});
