+++ packages/websocket/tests/trpc-timeline.test.ts
@@ -482,335 +482,4 @@ describe("timeline router", () => {
       await expect(caller.timeline.delete({ id: "seq-1" })).rejects.toThrow();
     });
   });
-
-  describe("versions", () => {
-    const docWithClip: TimelineDocument = {
-      tracks: [],
-      clips: [
-        {
-          id: "clip-1",
-          trackId: "v1",
-          name: "Clip",
-          startMs: 0,
-          durationMs: 1000,
-          mediaType: "video",
-          sourceType: "imported",
-          status: "generated",
-          locked: false,
-          versions: [
-            {
-              id: "v0",
-              createdAt: "2026-01-01T00:00:00Z",
-              jobId: "j",
-              assetId: "a",
-              workflowUpdatedAt: "2026-01-01T00:00:00Z",
-              dependencyHash: "h",
-              paramOverridesSnapshot: {},
-              status: "success"
-            }
-          ]
-        }
-      ],
-      markers: []
-    };
-
-    it("list returns the clip versions", async () => {
-      TS.findById.mockResolvedValue(
-        makeSeq({ document: JSON.stringify(docWithClip) })
-      );
-      const caller = createCaller(makeCtx());
-      const out = await caller.timeline.versions.list({
-        id: "seq-1",
-        clipId: "clip-1"
-      });
-      expect(out).toHaveLength(1);
-      expect(out[0]?.id).toBe("v0");
-    });
-
-    it("list 404s on unknown clip", async () => {
-      TS.findById.mockResolvedValue(
-        makeSeq({ document: JSON.stringify(docWithClip) })
-      );
-      const caller = createCaller(makeCtx());
-      await expect(
-        caller.timeline.versions.list({ id: "seq-1", clipId: "nope" })
-      ).rejects.toThrow();
-    });
-
-    it("append adds a new version and persists", async () => {
-      TS.findById.mockResolvedValue(
-        makeSeq({ document: JSON.stringify(docWithClip) })
-      );
-      TS.update.mockResolvedValue(undefined);
-      const caller = createCaller(makeCtx());
-      const out = await caller.timeline.versions.append({
-        id: "seq-1",
-        clipId: "clip-1",
-        jobId: "job-2",
-        assetId: "asset-2",
-        dependencyHash: "h2",
-        workflowUpdatedAt: "2026-01-02T00:00:00Z"
-      });
-      expect(out.id).toBe("version-id");
-      expect(out.status).toBe("success");
-      expect(TS.update).toHaveBeenCalled();
-      const updateArgs = TS.update.mock.calls[0]?.[1];
-      const persisted = JSON.parse(
-        updateArgs?.document as string
-      ) as TimelineDocument;
-      expect(persisted.clips[0].versions).toHaveLength(2);
-    });
-
-    it("append prunes successful versions beyond cap, keeping favorites", async () => {
-      // Build a clip with 10 successful versions (none favorited) + 1 favorite
-      const existingVersions = Array.from({ length: 10 }, (_, i) => ({
-        id: `v-${i}`,
-        createdAt: `2026-01-01T00:00:0${i}Z`,
-        jobId: `j-${i}`,
-        assetId: `a-${i}`,
-        workflowUpdatedAt: "2026-01-01T00:00:00Z",
-        dependencyHash: `h${i}`,
-        paramOverridesSnapshot: {},
-        status: "success" as const,
-        favorite: i === 0 // first one is favorited
-      }));
-
-      const docWithManyVersions: TimelineDocument = {
-        tracks: [],
-        clips: [
-          {
-            id: "clip-1",
-            trackId: "v1",
-            name: "Clip",
-            startMs: 0,
-            durationMs: 1000,
-            mediaType: "video",
-            sourceType: "imported",
-            status: "generated",
-            locked: false,
-            versions: existingVersions
-          }
-        ],
-        markers: []
-      };
-
-      TS.findById.mockResolvedValue(
-        makeSeq({ document: JSON.stringify(docWithManyVersions) })
-      );
-      TS.update.mockResolvedValue(undefined);
-      const caller = createCaller(makeCtx());
-      await caller.timeline.versions.append({
-        id: "seq-1",
-        clipId: "clip-1",
-        jobId: "job-new",
-        assetId: "asset-new",
-        dependencyHash: "h-new",
-        workflowUpdatedAt: "2026-01-02T00:00:00Z"
-      });
-      const updateArgs = TS.update.mock.calls[0]?.[1];
-      const persisted = JSON.parse(
-        updateArgs?.document as string
-      ) as TimelineDocument;
-      const versions = persisted.clips[0].versions!;
-      // Favorites are retained on top of the non-favorite cap (mirrors sketch):
-      // the 1 favorite plus the newest MAX_SUCCESSFUL_VERSIONS non-favorites.
-      // Here there are exactly 10 non-favorite successes (9 old + the new one),
-      // so all are kept and the favorite is extra: 11 total.
-      expect(versions.filter((v) => !v.favorite).length).toBeLessThanOrEqual(
-        10
-      );
-      // The favorite version must still be present.
-      expect(versions.some((v) => v.id === "v-0")).toBe(true);
-      // The just-appended version is retained.
-      expect(versions.some((v) => v.jobId === "job-new")).toBe(true);
-    });
-
-    it("append keeps the just-created version even when favorites fill the cap", async () => {
-      // Regression (#16): previously, 10 favorites zeroed the non-favorite
-      // slots and the freshly-appended version was silently dropped while still
-      // being returned to the client as saved. Favorites must not evict the new
-      // generation result.
-      const allFavVersions = Array.from({ length: 10 }, (_, i) => ({
-        id: `vf-${i}`,
-        createdAt: `2026-01-01T00:00:0${i}Z`,
-        jobId: `j-${i}`,
-        assetId: `a-${i}`,
-        workflowUpdatedAt: "2026-01-01T00:00:00Z",
-        dependencyHash: `h${i}`,
-        paramOverridesSnapshot: {},
-        status: "success" as const,
-        favorite: true
-      }));
-
-      const docAllFav: TimelineDocument = {
-        tracks: [],
-        clips: [
-          {
-            id: "clip-1",
-            trackId: "v1",
-            name: "Clip",
-            startMs: 0,
-            durationMs: 1000,
-            mediaType: "video",
-            sourceType: "imported",
-            status: "generated",
-            locked: false,
-            versions: allFavVersions
-          }
-        ],
-        markers: []
-      };
-
-      TS.findById.mockResolvedValue(
-        makeSeq({ document: JSON.stringify(docAllFav) })
-      );
-      TS.update.mockResolvedValue(undefined);
-      const caller = createCaller(makeCtx());
-      const out = await caller.timeline.versions.append({
-        id: "seq-1",
-        clipId: "clip-1",
-        jobId: "job-new",
-        assetId: "asset-new",
-        dependencyHash: "h-new",
-        workflowUpdatedAt: "2026-01-02T00:00:00Z"
-      });
-      const updateArgs = TS.update.mock.calls[0]?.[1];
-      const persisted = JSON.parse(
-        updateArgs?.document as string
-      ) as TimelineDocument;
-      const versions = persisted.clips[0].versions!;
-      // All 10 favorites survive (favorites are not counted against the cap).
-      expect(versions.filter((v) => v.favorite).length).toBe(10);
-      // The just-created non-favorite version is retained, not dropped.
-      expect(versions.some((v) => v.id === out.id)).toBe(true);
-      expect(versions.filter((v) => !v.favorite)).toHaveLength(1);
-      expect(versions.length).toBe(11);
-    });
-
-    it("setFavorite toggles the favorite flag", async () => {
-      TS.findById.mockResolvedValue(
-        makeSeq({ document: JSON.stringify(docWithClip) })
-      );
-      TS.update.mockResolvedValue(undefined);
-      const caller = createCaller(makeCtx());
-      const out = await caller.timeline.versions.setFavorite({
-        id: "seq-1",
-        clipId: "clip-1",
-        versionId: "v0",
-        favorite: true
-      });
-      expect(out.favorite).toBe(true);
-      expect(TS.update).toHaveBeenCalled();
-    });
-
-    it("setFavorite 404s on unknown version", async () => {
-      TS.findById.mockResolvedValue(
-        makeSeq({ document: JSON.stringify(docWithClip) })
-      );
-      const caller = createCaller(makeCtx());
-      await expect(
-        caller.timeline.versions.setFavorite({
-          id: "seq-1",
-          clipId: "clip-1",
-          versionId: "nope",
-          favorite: true
-        })
-      ).rejects.toThrow();
-    });
-
-    it("delete removes the version and persists", async () => {
-      TS.findById.mockResolvedValue(
-        makeSeq({ document: JSON.stringify(docWithClip) })
-      );
-      TS.update.mockResolvedValue(undefined);
-      const caller = createCaller(makeCtx());
-      const out = await caller.timeline.versions.delete({
-        id: "seq-1",
-        clipId: "clip-1",
-        versionId: "v0"
-      });
-      expect(out.ok).toBe(true);
-      const updateArgs = TS.update.mock.calls[0]?.[1];
-      const persisted = JSON.parse(
-        updateArgs?.document as string
-      ) as TimelineDocument;
-      expect(persisted.clips[0].versions).toHaveLength(0);
-    });
-
-    it("delete 404s on unknown version", async () => {
-      TS.findById.mockResolvedValue(
-        makeSeq({ document: JSON.stringify(docWithClip) })
-      );
-      const caller = createCaller(makeCtx());
-      await expect(
-        caller.timeline.versions.delete({
-          id: "seq-1",
-          clipId: "clip-1",
-          versionId: "nope"
-        })
-      ).rejects.toThrow();
-    });
-  });
-
-  describe("clips", () => {
-    const clipDoc: TimelineDocument = {
-      tracks: [],
-      clips: [
-        {
-          id: "clip-1",
-          trackId: "t1",
-          name: "Generated",
-          startMs: 500,
-          durationMs: 1000,
-          mediaType: "video",
-          sourceType: "generated",
-          workflowId: "wf-1",
-          paramOverrides: { prompt: "hello" },
-          status: "generated",
-          locked: false,
-          versions: []
-        }
-      ],
-      markers: []
-    };
-
-    it("delete removes only the clip", async () => {
-      TS.findById.mockResolvedValue(
-        makeSeq({ document: JSON.stringify(clipDoc) })
-      );
-      TS.update.mockResolvedValue(undefined);
-
-      const caller = createCaller(makeCtx());
-      const out = await caller.timeline.clips.delete({
-        id: "seq-1",
-        clipId: "clip-1"
-      });
-
-      expect(out.ok).toBe(true);
-      const updateArgs = TS.update.mock.calls[0]?.[1];
-      const persisted = JSON.parse(
-        updateArgs?.document as string
-      ) as TimelineDocument;
-      expect(persisted.clips).toHaveLength(0);
-    });
-
-    it("duplicate copies overrides and shares the workflow", async () => {
-      TS.findById.mockResolvedValue(
-        makeSeq({ document: JSON.stringify(clipDoc) })
-      );
-      TS.update.mockResolvedValue(undefined);
-
-      const caller = createCaller(makeCtx());
-      const out = await caller.timeline.clips.duplicate({
-        id: "seq-1",
-        clipId: "clip-1",
-        deltaMs: 2000
-      });
-
-      expect(out.id).toBe("version-id");
-      expect(out.workflowId).toBe("wf-1");
-      expect(out.startMs).toBe(2500);
-      expect(out.paramOverrides).toEqual({ prompt: "hello" });
-    });
-  });
 });
