+++ packages/websocket/tests/ws-phase2.test.ts
@@ -95,7 +95,6 @@ describe("T-WS-8: Username validation", () => {
 //
 // JSON ops (list, info) have moved to the tRPC `files` router.
 // Tests for those use the tRPC caller directly against a real temp dir.
-// The REST handler (/api/files/download) is tested for download + traversal.
 
 describe("T-WS-9: File browser API", () => {
   let tmpDir: string;
@@ -117,7 +116,7 @@ describe("T-WS-9: File browser API", () => {
     await fs.rm(tmpDir, { recursive: true, force: true });
   });
 
-  // ── list / info — now in tRPC files router ───────────────────────────
+  // ── list — now in tRPC files router ──────────────────────────────────
 
   it("tRPC files.list lists the home directory", async () => {
     // Uses real homedir — just assert shape (entries may vary per machine)
@@ -132,15 +131,6 @@ describe("T-WS-9: File browser API", () => {
     }
   });
 
-  it("tRPC files.info returns file metadata", async () => {
-    const caller = createCaller(makeFileCtx());
-    // info uses homedir as root — test with "." to get root info
-    const result = await caller.files.info({ path: "." });
-    expect(result).toHaveProperty("name");
-    expect(result).toHaveProperty("is_dir");
-    expect(typeof result.modified_at).toBe("string");
-  });
-
   it("tRPC files.list returns 404 for nonexistent directory", async () => {
     const caller = createCaller(makeFileCtx());
     await expect(
@@ -148,24 +138,10 @@ describe("T-WS-9: File browser API", () => {
     ).rejects.toMatchObject({ code: "NOT_FOUND" });
   });
 
-  it("tRPC files.info returns 404 for nonexistent file", async () => {
-    const caller = createCaller(makeFileCtx());
-    await expect(
-      caller.files.info({ path: "nonexistent-file-xyz.txt" })
-    ).rejects.toMatchObject({ code: "NOT_FOUND" });
-  });
-
   it("tRPC files.list rejects path traversal with FORBIDDEN", async () => {
     const caller = createCaller(makeFileCtx());
     await expect(
       caller.files.list({ path: "/../../../etc" })
-    ).rejects.toMatchObject({ code: "FORBIDDEN" });
-  });
-
-  it("tRPC files.info rejects path traversal with FORBIDDEN", async () => {
-    const caller = createCaller(makeFileCtx());
-    await expect(
-      caller.files.info({ path: "/../../etc/passwd" })
     ).rejects.toMatchObject({ code: "FORBIDDEN" });
   });
 
@@ -183,28 +159,6 @@ describe("T-WS-9: File browser API", () => {
       fileOpts
     );
     expect(res.status).toBe(404);
-  });
-
-  // ── download — stays as REST ─────────────────────────────────────────
-
-  it("GET /api/files/download?path=/hello.txt returns file content", async () => {
-    const res = await handleFileRequest(
-      new Request("http://localhost/api/files/download?path=/hello.txt"),
-      fileOpts
-    );
-    expect(res.status).toBe(200);
-    const text = await res.text();
-    expect(text).toBe("hello world");
-  });
-
-  it("GET /api/files/download with path traversal returns 403", async () => {
-    const res = await handleFileRequest(
-      new Request(
-        "http://localhost/api/files/download?path=/../../../etc/passwd"
-      ),
-      fileOpts
-    );
-    expect(res.status).toBe(403);
   });
 });
